+++ sqllin-driver/src/mingwMain/kotlin/com/ctrip/sqllin/driver/platform/UtilsMinGW.kt
@@ -2,6 +2,7 @@ package com.ctrip.sqllin.driver.platform
 
 import kotlinx.cinterop.ByteVar
 import kotlinx.cinterop.CPointer
+import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.toKStringFromUtf8
 
 /**
@@ -9,6 +10,7 @@ import kotlinx.cinterop.toKStringFromUtf8
  * @author yqiao
  */
 
+@OptIn(ExperimentalForeignApi::class)
 internal actual fun bytesToString(bv: CPointer<ByteVar>): String = bv.toKStringFromUtf8()
 
 internal actual inline val separatorChar: Char